Adding first version of data assessment
