+++ dvc.lock
@@ -71,13 +71,13 @@ stages:
     deps:
     - path: ./data/raw/gamelogs/
       hash: md5
-      md5: cc791945f5c7f644727956b4dfc1291a.dir
-      size: 6387190
+      md5: ab9a32ea1384c29cd29096ced2c6e3a2.dir
+      size: 6392194
       nfiles: 17
     - path: ./data/raw/schedules/
       hash: md5
-      md5: 071119142fb16ae1472aa3ef694b71a1.dir
-      size: 2882212
+      md5: af89928234e0d4071ef59cf230107057.dir
+      size: 2882808
       nfiles: 17
     - path: src/gamelog_schedule_unification.py
       hash: md5
@@ -97,8 +97,8 @@ stages:
     outs:
     - path: ./data/input/nba_gamelog_schedule_dataset.csv
       hash: md5
-      md5: 8ce94e24444dec9d85cc5616555e3705
-      size: 7180452
+      md5: b502cc798473b44917080506e2e03222
+      size: 7186030
   player_attributes_salaries_unification:
     cmd: python3 -m src.transform.player_attributes_salaries_unification --config=params.yaml
     deps:
@@ -122,13 +122,13 @@ stages:
     outs:
     - path: ./data/raw/gamelogs/
       hash: md5
-      md5: cc791945f5c7f644727956b4dfc1291a.dir
-      size: 6387190
+      md5: ab9a32ea1384c29cd29096ced2c6e3a2.dir
+      size: 6392194
       nfiles: 17
     - path: ./data/raw/schedules/
       hash: md5
-      md5: 071119142fb16ae1472aa3ef694b71a1.dir
-      size: 2882212
+      md5: af89928234e0d4071ef59cf230107057.dir
+      size: 2882808
       nfiles: 17
   pre_cleaning_dataset:
     cmd:
@@ -137,8 +137,8 @@ stages:
     deps:
     - path: ./data/input/nba_gamelog_schedule_dataset.csv
       hash: md5
-      md5: 8ce94e24444dec9d85cc5616555e3705
-      size: 7180452
+      md5: b502cc798473b44917080506e2e03222
+      size: 7186030
     - path: src/pre_cleaning_dataset.py
       hash: md5
       md5: 3bb713a147b28a797ca42bb61561df5e
@@ -150,8 +150,8 @@ stages:
     outs:
     - path: ./data/processed/nba_games_training_dataset_pre_cleaned.csv
       hash: md5
-      md5: 542edae565d1b433ed2ad319d1d082eb
-      size: 2751100
+      md5: 89bb145c5cb0b5afe6cb5f13da70df48
+      size: 2753226
   previous_games_average_features:
     cmd: python3 -m src.previous_games_average_features --config=params.yaml
     deps:
@@ -209,8 +209,8 @@ stages:
     deps:
     - path: ./data/processed/nba_games_training_dataset_pre_cleaned.csv
       hash: md5
-      md5: 542edae565d1b433ed2ad319d1d082eb
-      size: 2751100
+      md5: 89bb145c5cb0b5afe6cb5f13da70df48
+      size: 2753226
     - path: src/features_engineering.py
       hash: md5
       md5: 84fc9d393183c99ae9a0d48b35d537ca
@@ -222,8 +222,8 @@ stages:
     outs:
     - path: ./data/processed/nba_games_training_dataset_cleaned_w_features.csv
       hash: md5
-      md5: b43401878e8810d561097e4df1232774
-      size: 15524699
+      md5: d286a3ba7a6a46722a972e249ffeaff1
+      size: 15109284
   get_y_variables:
     cmd:
     - python3 -m src.get_y_variables --input-file-folder-name ./data/processed/nba_games_training_dataset_cleaned_w_features.csv
@@ -231,8 +231,8 @@ stages:
     deps:
     - path: ./data/processed/nba_games_training_dataset_cleaned_w_features.csv
       hash: md5
-      md5: b43401878e8810d561097e4df1232774
-      size: 15524699
+      md5: d286a3ba7a6a46722a972e249ffeaff1
+      size: 15109284
     - path: src/get_y_variables.py
       hash: md5
       md5: 2bb152cd5c91f8264b681b480b54eb21
@@ -247,8 +247,8 @@ stages:
     outs:
     - path: ./data/processed/nba_games_training_dataset_final.csv
       hash: md5
-      md5: 9f5cf92f1944b1ca71db3488f7675357
-      size: 15799260
+      md5: bbaf78bfa6c530a43882824e2ce347ef
+      size: 15384027
   load_training_dataset_to_db:
     cmd: python3 -m src.load_training_dataset_to_db --config=params.yaml
     deps:
@@ -268,36 +268,44 @@ stages:
           nba_games_training_dataset: nba_games_training_dataset
           player_attributes_salaries_dataset: player_attributes_salaries_dataset
   post_cleaning_dataset:
-    cmd: python3 -m src.post_cleaning_dataset --config=params.yaml
+    cmd:
+    - python3 -m src.post_cleaning_dataset --input-file-folder-name ./data/processed/nba_games_training_dataset_final.csv
+      --output-file-folder-name ./data/output/nba_games_training_dataset_final.csv
     deps:
     - path: ./data/processed/nba_games_training_dataset_final.csv
       hash: md5
-      md5: 9f5cf92f1944b1ca71db3488f7675357
-      size: 15799260
+      md5: bbaf78bfa6c530a43882824e2ce347ef
+      size: 15384027
     - path: src/post_cleaning_dataset.py
       hash: md5
-      md5: f8efd0bcd3da17902eec1be1428b12f9
-      size: 1854
+      md5: 04448b56daf107e82704628589f19f5b
+      size: 3297
     params:
       params.yaml:
         base:
           log_level: INFO
+        get_y_variables:
+          input_file: data/processed/nba_games_training_dataset_cleaned_w_features
+          output_file: data/processed/nba_games_training_dataset_final
+        post_cleaning_dataset:
+          output_folder: data/output
+          output_file_name: nba_games_training_dataset_final
     outs:
     - path: ./data/output/nba_games_training_dataset_final.csv
       hash: md5
-      md5: 56d0a7a87f5ae7bc2197bedb4a677b06
-      size: 14080665
+      md5: 95a272ed6c99ad449be3b126bbc5a437
+      size: 13669023
   training_dataset_assessment:
     cmd: python3 -m src.training_dataset_assessment --config=params.yaml
     deps:
     - path: ./data/output/nba_games_training_dataset_final.csv
       hash: md5
-      md5: 56d0a7a87f5ae7bc2197bedb4a677b06
-      size: 14080665
+      md5: 95a272ed6c99ad449be3b126bbc5a437
+      size: 13669023
     - path: src/training_dataset_assessment.py
       hash: md5
-      md5: 2b3e6d025405e5fb4d76fe35c25d55fc
-      size: 38804
+      md5: b464ad2ac50b368fe86decb4b9d1792a
+      size: 6430
     params:
       params.yaml:
         base:
@@ -324,5 +332,5 @@ stages:
     deps:
     - path: ./data/output/nba_games_training_dataset_final.csv
       hash: md5
-      md5: 56d0a7a87f5ae7bc2197bedb4a677b06
-      size: 14080665
+      md5: 95a272ed6c99ad449be3b126bbc5a437
+      size: 13669023
