+++ src/post_cleaning_dataset.py
@@ -6,30 +6,30 @@ import argparse
 from pathlib import Path
 import pandas as pd
 import yaml
+import os
 
 from src.utils.logs import get_logger
 
+logger = get_logger(
+    "POST_CLEANED_DATASET", log_level='INFO'
+)
 
-def post_cleaning_dataset(config_path: Path) -> pd.DataFrame:
+def post_cleaning_dataset(
+        input_file_folder_name: str = 'data/processed/nba_games_training_dataset_final.csv',
+        output_file_folder_name: str ='data/output/nba_games_training_dataset_final.csv'
+        ) -> None:
     """
-    Load and pre-clean the NBA games training dataset.
+    Post Cleaning Dataset.
 
     Args:
-        config_path (Text): Path to the configuration file.
-
-    Returns:
-        pd.DataFrame: The pre-cleaned training dataset.
+        input_file_folder_name (str): Unified gamelogs and schedules for multiple seasons
+        output_file_folder_name (str): Path where to save the cleaned unified dataframe.
     """
-    with open(config_path, encoding="utf-8") as conf_file:
-        config_params = yaml.safe_load(conf_file)
-
-    logger = get_logger(
-        "POST_CLEANED_DATASET", log_level=config_params["base"]["log_level"]
-    )
 
     nba_games_training_dataset = pd.read_csv(
-        "./data/processed/nba_games_training_dataset_final.csv"
-    )
+        input_file_folder_name
+        )
+
     logger.info("Shape of the DataFrame %s", str(nba_games_training_dataset.shape))
 
     # Column Selection
@@ -50,14 +50,78 @@ def post_cleaning_dataset(config_path: Path) -> pd.DataFrame:
     
     logger.info("Shape of the DataFrame %s", str(nba_games_training_dataset.shape))
 
+
     nba_games_training_dataset.to_csv(
-        "./data/output/nba_games_training_dataset_final.csv", index=False
+        output_file_folder_name, index=False
     )
 
     logger.info("Post Cleaned NBA games data step complete")
 
+
+def get_args():
+    """
+    Parse command line arguments and return the parsed arguments.
+
+    Returns:
+        argparse.Namespace: Parsed command line arguments.
+    """
+    _dir = Path(__file__).parent.resolve()
+    parser = argparse.ArgumentParser()
+
+    parser.add_argument(
+        "--params-file",
+        type=Path,
+        default="params.yaml",
+    )
+
+    args, _ = parser.parse_known_args()
+    params = yaml.safe_load(args.params_file.open())
+
+    get_y_variables_params = params["get_y_variables"]
+    post_cleaning_dataset_params = params["post_cleaning_dataset"]
+
+    output_file_folder_name = os.path.join(
+        post_cleaning_dataset_params['output_folder'], 
+        post_cleaning_dataset_params['output_file_name'] +  ".csv"
+        )
+    
+    parser.add_argument(
+        "--input-file-folder-name",
+        dest="input_file_folder_name",
+        type=str,
+        default=get_y_variables_params["output_file"],
+    )
+    
+    parser.add_argument(
+        "--output-file-folder-name",
+        dest="output_file_folder_name",
+        type=Path,
+        default=output_file_folder_name,
+    )
+
+    parser.add_argument(
+        "--output-folder",
+        dest="output_folder",
+        type=Path,
+        default=post_cleaning_dataset_params["output_folder"],
+    )
+
+    args = parser.parse_args()
+
+    args.output_folder.parent.mkdir(
+        parents=True, 
+        exist_ok=True)
+        
+    return args
+
+def main():
+    """Run the Post Cleaning Dataset Step."""
+    args = get_args()
+
+    post_cleaning_dataset(
+        input_file_folder_name=args.input_file_folder_name,
+        output_file_folder_name=args.output_file_folder_name
+    )
+
 if __name__ == "__main__":
-    arg_parser = argparse.ArgumentParser()
-    arg_parser.add_argument("--config-params", dest="config_params", required=True)
-    args = arg_parser.parse_args()
-    post_cleaning_dataset(config_path=args.config_params)
+    main()
